requirements.txt
ultralytics
torch
scipy
matplotlib
pillow
streamlit